+++ app/(instagramer)/wallet/page.tsx
@@ -1,21 +1,5 @@
-"use client";
-
-import { useSession } from "next-auth/react";
-import { useRouter } from "next/navigation";
-import { useEffect } from "react";
+import { redirect } from "next/navigation";
 
 export default function Page() {
-  const router = useRouter();
-  const { data: session } = useSession({
-    required: true,
-    onUnauthenticated() {
-      router.push("/");
-    },
-  });
-
-  useEffect(() => {
-    router.push("/wallet/statistics");
-  }, [router]);
-
-  return null;
+  redirect("/wallet/payment");
 }