+++ i18n/az.ts
@@ -616,7 +616,7 @@ export default {
     subadmininvitation: "Əməkdaş dəvəti",
     subadmininvitationexplain: "Instagram idarəetmə dəvəti əməkdaş olaraq",
     CollaboratorExplain: "Post və ya Reels-i Bir Neçə Instagram Səhifəsi Arasında Birgə Paylaşın",
-
+    Renew_Subscription: "Abunəliyi yenilə",
     upgrade: "Yüksəlt",
     Deactivate: "Deaktiv et",
     done: "Tamamlandı",
@@ -3071,6 +3071,8 @@ export default {
     Notify_InvalidItemId: "Maddə ID-si etibarlı deyil",
     Notify_IpInvalid: "Daxil edilmiş IP etibarlı deyil",
     Notify_InstagramRedirectInTenSeconds: "10 saniyə ərzində avtomatik olaraq Instagram-a yönləndiriləcəksiniz",
+    Notify_InstagramRedirect: "Instagram-a yönləndirilir",
+    Notify_InstagramRedirectExplain: "IP ünvanınız etibarlı olmadığı üçün VPN-in aktiv olduğundan əmin olun",
     Notify_ThisInstagramAccountAddedBefore: "Bu Instagram hesabı əvvəllər əlavə edilib",
     Notify_InvalidTicketId: "Bilet ID-si etibarlı deyil",
     Notify_ExpiredSignature: "İmza müddəti bitib",